+++ dbengine/nr_kernel/nr_am/src/nram_storage/rocks_handler.h
@@ -12,5 +12,5 @@ bool RocksClientIndexPut(NRIndexKey ikey, NRIndexValue ivalue);
 bool RocksClientIndexDelete(NRIndexKey ikey);
 bool RocksClientIndexRangeScan(NRIndexKey start_key, NRIndexKey end_key, NRIndexKey **out_keys, NRIndexValue **out_results, int *out_count);
 
-/* Bulk load for efficient index building */
-bool RocksClientIndexBulkLoad(Oid indexOid, int32 *keys, uint64 *values, int count);
+/* Bulk load for efficient index building (supports INT and BIGINT) */
+bool RocksClientIndexBulkLoad(Oid indexOid, int64 *keys, uint64 *values, int count);
